package_rpm.bbclass: Don't pass a 4th parameter to rpm target
rpm --target expects 3 parameters, split by a dash:
	arch-vendor-os

However, glibc may add another parameter, corresponding to
the extension used. This is actually common on ARM archs.
However, rpm 4.1.4 is not prepared to handle it. When 4
parameters are found, it will use the last one as the OS,
causing the image build to fail.

Fix this by always discarding the 4th parameter, if present.

This patch is a co-work with Thiago Santos <thiagoss@osg.samsung.com>,
with actually come with the final solution.

Change-Id: I8ab3473ef4e6ff1501a1f82db0ab2bb8d7d9df50
Signed-off-by: Thiago Santos <thiagoss@osg.samsung.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>