Remove -j16 from make
There's no need to restrict/enforce parallel make to 16. This should
be limited by ${PARALLEL_MAKE}.

In the case of the machine I'm using, there are 32 CPU's. So,
PARALLEL_MAKE is set, by default, to 32. With such config,
it gets this warning:

	NOTE: make -j 32 -j16

On several packages.

As the ${PARALLEL_MAKE} parameter is already passed to make,
just removing -j16 seem to be the right thing to do here.

Change-Id: I67bbbedf89f79b0712503e45e449b109a8e53d2c
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>