meta-tizen: Ignore .gitreview
The .gitreview is the configuration file for gerrit git-review
command. As the tizen-distro layer merge scripts check if the
meta-tizen tree is dirty, we need to ignore it, or otherwise
it will refuse to merge from it.

Change-Id: Ib2d41b69b4d956392176ce98b8625dac6ed8024e
(From meta-tizen rev: af886b35e3f715a28d17d08e7f54ab613444ff37)

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>